Added extra copy_term to lexical lookup
The extra copy_term will prevent errors caused by accidental variable
reuse between the prosodic term and the semantic term. I’m assuming
here that nobody will ever consciously want this sharing and that it is
easy to make errors by using lamda(Z,…) in the prosodics and
lambda(Z,…) in the semantics.